audioInstance-methods {audio}	R Documentation

Audio instance class methods

Description

audioInstances supports most methods relevant to them. See the corresponding generics help for details.

Noteworthy is that $data is the canonical way to get data associated with an audio instance - i.e. played or recorded content.

Usage

## S3 method for class 'audioInstance'
x$name
## S3 method for class 'audioInstance'
resume(x, ...)
## S3 method for class 'audioInstance'
pause(x, ...)
## S3 method for class 'audioInstance'
rewind(x, ...)
## S3 method for class 'audioInstance'
close(con, ...)
## S3 method for class 'audioInstance'
play(x, ...)
## S3 method for class 'audioInstance'
print(x, ...)

Arguments

x
audio instance

name
name of the attribute - currently only "data" is supported

con
audio instance (the name is unfortuantely defined in the close generic like this)

...
ignored

[Package audio version 0.1-7 Index]