audioSample-methods {audio}	R Documentation

Audio sample class methods

Description

audioSample methods behave in the same way as the underlying methods of numeric vectors and matrices except that they preserve the attributes and class of the objects.

Usage

  ## S3 method for class 'audioSample'
x$name
  ## S3 replacement method for class 'audioSample'
x$name <- value
  ## S3 method for class 'audioSample'
x[..., drop = FALSE]
  ## S3 method for class 'audioSample'
as.Sample(x, ...) 
  ## S3 method for class 'audioSample'
print(x, ...)

Arguments

x
sample object

name
name of the attribute to get/set

value
value to set

drop
see `[` operator documentation

...
parameters passed to the object-specific method

Examples

x <- audioSample(sin(1:8000/10), 8000)
x$rate
x[1:10]

[Package audio version 0.1-7 Index]