controls {audio}	R Documentation

Control audio instance

Description

pause pauses (stops) audio recording or playback

rewind rewinds audio recording or playback, i.e., makes the beginning of the source (or target) object the current audio position.

resume resumes previously paused audio recording or playback

Usage

pause(x, ...)
rewind(x, ...)
resume(x, ...)

Arguments

x
instance object

...
optional arguments passed to the method specific to the object

Value

All functions return TRUE on success and FALSE on failure. All methods are generics and intended to apply to similar asynchronous operations (e.g. movie playback etc.).

See Also

play, record

[Package audio version 0.1-7 Index]