play {audio}	R Documentation

Play audio

Description

play plays audio

Usage

play(x, ...)
## S3 method for class 'audioSample'
play(x, rate, ...)
## S3 method for class 'Sample'
play(x, ...) 
## Default S3 method:
play(x, rate = 44100, ...)

Arguments

x
data to play

rate
sample rate - it is inferred from the object (where possible) if not specified

...
optional arguments passed to the method specific to the object being played

Value

Returns an audio instance object which can be used to control the playback subsequently.

Examples

play(sin(1:10000/20))

[Package audio version 0.1-7 Index]