wave {audio}	R Documentation

WAVE file manipulations

Description

load.wave loads a sample from a WAVE file

save.wave saves a sample into a WAVE file

Usage

load.wave(where)
save.wave(what, where)

Arguments

where
file name of the file to load from or save to

what
audioSample object to save

Details

WAVE is a RIFF (Resource Interchange File Format) widely used for storage of uncompressed audio data. It is often identified by the extension .WAV on DOS-legacy systems (such as Windows). Although WAVE files may contain compressed data, the above functions only support plain, uncompressed PCM data.

Value

load.wave returns an object of the class audioSample as loaded from the WAVE file

save.wave always returns NULL

See Also

audioSample, play, record

[Package audio version 0.1-7 Index]